requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Faker==8.12.1
kombu==5.1.0
phonenumbers==8.12.31
Pillow==8.3.1
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
